requirements.txt
flask
python-dotenv
pymongo
dnspython
bcrypt
flask-cors
pyjwt